last commit

+++ src/main/java/agents/Attaquant.java
@@ -30,7 +30,7 @@ public class Attaquant extends Defenseur implements IUniteLibre {
     /**
      * Constructor.
      */
-    public Attaquant(Base b, int pv, int pa, int poa, double pov, Case<Point> c, Carte ca) {
+    public Attaquant(Base b, int pv, int pa, double poa, double pov, Case<Point> c, Carte ca) {
     // Start of user code for constructor Unite
     super(b, pv, pa, poa, pov, c, ca);
     // End of user code
